issue_project_fixed/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
httpx==0.28.1
tenacity==8.2.2
rich==13.5.0
click==8.1.7
typing-extensions==4.8.0
pytest==8.4.2
